Daily Upload
Changed the rain to be a particle system instead of a load of instances
Moved around some objects to be infront of and behind the particles
Added try catch around decrypting db. If db wasnt encrypted it also wont
be decrypted
Added new scene load method so scenes can be loaded addative or not.
Added snow effect with particles
Added thunder effect with flas and thunder in clouds